Added Kp=0 special case management, fixed fault diagnostic mismatch.

+++ emBODY/eBcode/arch-arm/embobj/plus/mc/EOspeedmeter_hid.h
@@ -57,7 +57,7 @@ struct EOabsCalibratedEncoder_hid
     int32_t delta;
 
     uint8_t  ID;
-    uint8_t  state_mask;
+    uint16_t  state_mask;
     uint8_t  first_valid_data;
     uint16_t invalid_fault_cnt;
     uint16_t timeout_fault_cnt;
